create separate controller
