+++ call-center/server/controllers/calls.controller.ts
@@ -1,0 +1,34 @@
+import { Request, Response } from 'express';
+import { FindManyOptions, getManager } from 'typeorm';
+import { CallLeg } from '../entities/callLeg.entity';
+
+class CallsController {
+  public static get = async function (req: Request, res: Response) {
+    let { limit, ...callLegQuery } = req.query;
+    let findOpts = {
+      // NOTE You'll likely want to do some validation here
+      // to check for valid columns to query
+      where: callLegQuery,
+    } as FindManyOptions;
+
+    if (limit) {
+      findOpts.take = parseInt(limit as string);
+    }
+
+    try {
+      let callLegRepository = getManager().getRepository(CallLeg);
+
+      res.json({
+        calls: await callLegRepository.find(findOpts),
+      });
+    } catch (e) {
+      console.error(e);
+
+      res
+        .status(e && e.name === 'EntityNotFound' ? 404 : 500)
+        .send({ error: e.toString() });
+    }
+  };
+}
+
+export default CallsController;
