+++ call-center/server/routes/callControl.ts
@@ -3,8 +3,6 @@ import CallControlController from '../controllers/callControl.controller';
 
 let router = express.Router();
 
-router.get('/', CallControlController.get);
-
 // Actions
 router.post('/actions/dial', CallControlController.dial);
 router.post('/actions/conferences/invite', CallControlController.invite);
